requirements.txt
Pillow==9.5.0
Requests==2.30.0
streamlit==1.31.0
streamlit_lottie==0.0.3
streamlit_timeline==0.0.2
llama-index>=0.8.5
llama-cpp-python==0.1.9
nltk==3.8.1
sentence_transformers
langchain
beautifulsoup4
streamlit-carousel
st-clickable-images
